Made delete meeting work again
